# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Claw-Empire API",
    "version": "1.2.4",
    "description": "Baseline OpenAPI contract for commonly used Claw-Empire endpoints."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current origin (recommended for local docs)"
    },
    {
      "url": "http://127.0.0.1:8790",
      "description": "Local development (loopback)"
    },
    {
      "url": "http://localhost:8790",
      "description": "Local development (localhost)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API_AUTH_TOKEN"
      },
      "inboxSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-inbox-secret"
      }
    },
...
  },
  "paths": {
...
    "/api/messenger/sessions": {
      "get": {
        "tags": ["runtime"],
        "summary": "List messenger runtime sessions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Messenger runtime sessions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "sessions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "sessionKey": {
                            "type": "string"
                          },
                          "channel": {
                            "type": "string"
                          },
                          "targetId": {
                            "type": "string"
                          },
                          "enabled": {
                            "type": "boolean"
                          },
                          "displayName": {
                            "type": "string"
                          }
                        },
                        "required": ["sessionKey", "channel", "targetId", "enabled", "displayName"],
                        "additionalProperties": true
                      }
                    }
                  },
                  "required": ["ok", "sessions"],
                  "additionalProperties": true
                },
                "example": {
                  "ok": true,
                  "sessions": [
                    {
                      "sessionKey": "telegram:ops",
                      "channel": "telegram",
                      "targetId": "-1001234567890",
                      "enabled": true,
                      "displayName": "Ops"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": true
                },
                "example": {
                  "error": "unauthorized",
                  "message": "Authentication is required."
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": true
                },
                "example": {
                  "error": "forbidden",
                  "message": "You do not have permission to access this resource."
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": true
                },
                "example": {
                  "error": "too_many_requests",
                  "message": "Rate limit exceeded."
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": true
                },
                "example": {
                  "error": "internal_error",
                  "message": "Unexpected server error."
                }
              }
            }
          }
        }
      }
    },
    "/api/messenger/receiver/telegram": {
      "get": {
        "tags": ["runtime"],
        "summary": "Get telegram receiver status",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Telegram receiver status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "status": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  },
                  "required": ["ok", "status"],
                  "additionalProperties": true
                },
                "example": {
                  "ok": true,
                  "status": {
                    "running": true,
                    "mode": "polling"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": true
                },
                "example": {
                  "error": "unauthorized",
                  "message": "Authentication is required."
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": true
                },
                "example": {
                  "error": "forbidden",
                  "message": "You do not have permission to access this resource."
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": true
                },
                "example": {
                  "error": "too_many_requests",
                  "message": "Rate limit exceeded."
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": true
                },
                "example": {
                  "error": "internal_error",
                  "message": "Unexpected server error."
                }
              }
            }
          }
        }
      }
    },
...
  }
}